chore: ditch the command bus layer
